refactored markdown to separate assembly
