+++ Portal/src/Datahub.Core/Datahub.Core.csproj
@@ -93,6 +93,7 @@
 		<GitHookFiles Include="../.githooks/*.*"></GitHookFiles>
 	</ItemGroup>
 	<ItemGroup>
+		<ProjectReference Include="..\..\..\Shared\src\Datahub.Markdown\Datahub.Markdown.csproj" />
 		<ProjectReference Include="..\Datahub.CatalogSearch\Datahub.CatalogSearch.csproj" />
 		<ProjectReference Include="..\Datahub.Metadata\Datahub.Metadata.csproj" />
 	</ItemGroup>
