+++ Portal/src/Datahub.Core/Services/Wiki/IWikiService.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Datahub.Markdown.Model;
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
