+++ Portal/src/Datahub.Portal/Components/EmbeddedWiki.razor
@@ -1,7 +1,7 @@
 ﻿@*EmbeddedWiki.razor*@
 
-@using Datahub.Core.Services.Docs
-@using Datahub.Portal.Pages.Resources
+@    using Datahub.Core.Services.Docs
+@    using Datahub.Portal.Pages.Resources
 @inject DocumentationService _docService
 @inject CultureService _cultureService
