+++ Portal/src/Datahub.Portal/Pages/Resources/ResourcesListView.razor
@@ -1,5 +1,5 @@
-@using Datahub.Core.Services.Docs
-@using MudBlazor.Utilities
+@    using Datahub.Core.Services.Docs
+@    using MudBlazor.Utilities
 
 <MudGrid>
     @foreach (var documentItem in DocumentItems)
