+++ Portal/src/Datahub.Portal/Pages/Resources/ResourcesMarkdown.razor
@@ -1,8 +1,8 @@
 @*ResourcesMarkdown.razor*@
 
-@using Datahub.Core.Services.Docs
-@using Datahub.Core.Model.UserTracking
-@using System.Text.RegularExpressions
+@    using Datahub.Core.Services.Docs
+@    using Datahub.Core.Model.UserTracking
+@    using System.Text.RegularExpressions
 @inject DocumentationService _docService
 @inject UserLocationManagerService UserLocationManagerService
 <MudGrid Spacing="3">
