+++ Portal/src/Datahub.Portal/Pages/Resources/ResourcesPage.razor
@@ -1,6 +1,6 @@
-﻿@using Datahub.Core.Services.Docs
-@using MudBlazor.Utilities
-@using System.Text.RegularExpressions
+﻿@    using Datahub.Core.Services.Docs
+@    using MudBlazor.Utilities
+@    using System.Text.RegularExpressions
 @inject DocumentationService _docService
 @inject IDbContextFactory<DatahubProjectDBContext> _dbContextFactory
 @inject CultureService _cultureService
