+++ Portal/src/Datahub.Portal/Pages/Resources/ResourcesPreviewCard.razor
@@ -1,5 +1,5 @@
-@using Datahub.Core.Services.Docs
-@using MudBlazor.Utilities
+@    using Datahub.Core.Services.Docs
+@    using MudBlazor.Utilities
 
 <MudPaper Elevation="0" Outlined Class="py-2 px-3" Style="@_cardStyle">
     <MudStack Justify="Justify.SpaceBetween" Style="height: 100%;">
