+++ Shared/src/Datahub.Markdown/Datahub.Markdown.csproj
@@ -1,0 +1,12 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+	<PropertyGroup>
+		<TargetFramework>net7.0</TargetFramework>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Nullable>enable</Nullable>
+	</PropertyGroup>
+	<ItemGroup>
+		<PackageReference Include="Markdig" Version="0.33.*" />
+		<PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="7.0.*" />
+	</ItemGroup>
+</Project>
